refactor: 훅 이름 변경
